+++ PowerSync/PowerSync.Common/Utils/ICloseable.cs
@@ -1,0 +1,12 @@
+namespace PowerSync.Common.Utils;
+
+public interface ICloseable
+{
+    public void Close();
+}
+
+public interface ICloseableAsync
+{
+    public Task Close();
+}
+