+++ PowerSync/PowerSync.Common/CHANGELOG.md
@@ -2,6 +2,49 @@
 
 ## 0.0.11-alpha.1
 
+- `StatusUpdated` and `StatusChanged` now both emit `SyncStatus` objects instead of just `StatusChanged`.
+- Converted most instances of a class inheriting from `EventStream<T>` into a class with an `EventManager` property called `Events`. This allows for subscribing to individual events instead of subscribing to all events and then filtering events manually.
+
+```csharp
+// Old
+var listener = db.Listen(cts.Token);
+foreach (PowerSyncDBEvent update in listener)
+{
+    // Manually filter updates
+    if (update.StatusChanged != null)
+    {
+        Console.WriteLine("status changed: " + update.StatusChanged!);
+    }
+}
+
+// Old (async)
+var listener = db.ListenAsync(cts.Token);
+await foreach (PowerSyncDBEvent update in listener)
+{
+    // Manually filter updates
+    if (update.StatusChanged != null)
+    {
+        Console.WriteLine("status changed: " + update.StatusChanged!);
+    }
+}
+
+// New
+var listener = db.Events.OnStatusChanged.Listen(cts.Token);
+await foreach (PowerSyncDBEvents.StatusChanged update in listener)
+{
+    // Events are filtered inherently
+    Console.WriteLine("status changed: " + update.Status);
+}
+
+// New (async) - recommended for most use cases
+var listener = db.Events.OnStatusChanged.ListenAsync(cts.Token);
+await foreach (PowerSyncDBEvents.StatusChanged update in listener)
+{
+    // Events are filtered inherently
+    Console.WriteLine("status changed: " + update.Status);
+}
+```
+
 - Pool read connections in `MDSQLiteAdapter`, improving performance in any case where multiple queries run simultaneously (eg. via `Watch`). The number of connections can be set via `MDSQLiteOptions.ReadPoolSize` and defaults to 5.
 - Updated to the latest version (0.4.11) of the core extension.
 - `MDSQLiteConnection` now runs query operations on another thread, which stops the caller thread from blocking.